requirements.txt
langgraph
langchain
langchain-openai
docker
python-dotenv
requests
pydantic
flask
flask-socketio
prometheus-client
